Invalidate server query cache
I've added an invalidate function to `WithServerQuery.js`, which can be used as e.g. `MyDomain.get.invalidate({ attr: 'key })`. It'll remove the server query from the `serverQueries` object against the key. It then returns the a fresh version of the query.

Anything naive about this approach? Works for me so far.